% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedFolds.R
\name{balancedFolds}
\alias{balancedFolds}
\title{Function producing stratified/ balanced folds for cross validation}
\usage{
balancedFolds(class.column.factor, cross.outer)
}
\description{
Get balanced folds for cross validation, which are used for tuning penalization parameters. This function is mainly used within the function \code{epsgo}. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
