% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_intsearch.R
\name{coef.sum.intsearch}
\alias{coef.sum.intsearch}
\title{Get coefficients for a model}
\usage{
coef.sum.intsearch(object, ...)
}
\description{
Get coefficients for a model after applying interval search for tuning parameters. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
