% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpImprovement.R
\name{ExpImprovement}
\alias{ExpImprovement}
\title{ExpImprovement}
\usage{
ExpImprovement(x_new, fmin, fit.gp, muX = NULL, muY = NULL,
  EI.eps = 0.01)
}
\description{
Calculates the expected improvement of the funtion \code{Q.func}. See the \code{R} package  \pkg{penalizedSVM} for details.
}
\details{
IPFStructPenalty
}
