% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheight.R
\name{mheight}
\alias{mheight}
\title{Structured penalized regression}
\usage{
mheight(y, foldid = NULL)
}
\arguments{
\item{y}{response matrix}

\item{foldid}{an optional vector identifying what fold each observation is in. The \code{mheigth} returns the maximum of the leave-one-fold data's minimun height. Default is \code{NULL}.}
}
\description{
Subfunction for the tree-lasso and IPF-tree-lass methods to compute the minimun height of the dendrogram.
}
\details{
IPFStructPenalty
}
