% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef_glmnet.R
\name{Plot.coef.glmnet}
\alias{Plot.coef.glmnet}
\title{function to highlight the path of a pre-specified set of variables within the coefficient path}
\usage{
Plot.coef.glmnet(cvfit, betas)
}
\description{
Creates several plots showing the coefficient path for the final model of a \code{cv.glmnet} fit and highlights the path of a pre-specified set of variables within the coefficient path. See the \code{R} package  \pkg{penalizedSVM} for details.
}
\details{
IPFStructPenalty
}
