% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_inter_search.R
\name{plot.sum.intsearch}
\alias{plot.sum.intsearch}
\title{Plot Summary object for interval search models}
\usage{
\method{plot}{sum.intsearch}(x, type = "summary", startN = 21, ...)
}
\description{
Produces a plot for summary object of a fitted interval search model. Plot 'visited' points against iteration steps. start.N points are initial points selected before interval search starts. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
