% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityselection.R
\name{stabpath}
\alias{stabpath}
\title{Stability path for glmnet models}
\usage{
stabpath(y, x, size = 0.632, steps = 100, weakness = 1,
  mc.cores = getOption("mc.cores", 2L), ...)
}
\description{
he function calculates the stability path for glmnet models, e.g. the selection probabilities of the features along the range of regularization parameters. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
