% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_clogit_interval.R
\name{tune.clogit.interval}
\alias{tune.clogit.interval}
\title{Wrapper function for conditional logistic lasso objects.}
\usage{
tune.clogit.interval(parms, x = x, y = y, x_test = NULL,
  y_test = NULL, strata = NULL, lambda = NULL, nfolds = 3,
  foldid = NULL, seed = 12345, standardize.response = FALSE,
  p = NULL, parallel = FALSE, verbose = FALSE, search.path = FALSE,
  ...)
}
\description{
Wrapper function for conditional logistic lasso objects used by epsgo function. This function is mainly used within the function \code{epsgo}. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
