% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_glmnet_interval.R
\name{tune.glmnet.interval}
\alias{tune.glmnet.interval}
\title{Wrapper function for glmnet objects.}
\usage{
tune.glmnet.interval(parms, x, y, weights, x_test = NULL,
  y_test = NULL, offset = NULL, lambda = NULL, num.nonpen = 0,
  nfolds = 10, type.measure = c("mse", "deviance", "class", "auc",
  "mae"), seed = 12345, grouped = TRUE, type.min = c("lambda.min",
  "lambda.1se"), family = "gaussian", foldid = NULL,
  intercept = TRUE, standardize.response = FALSE, p = NULL,
  parallel = FALSE, verbose = FALSE, search.path = FALSE,
  lib.loc = NULL, ...)
}
\description{
Wrapper function for glmnet objects used by epsgo function. This function is mainly used within the function \code{epsgo}. See the \code{R} package \pkg{c060} for details.
}
\details{
IPFStructPenalty
}
