% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_tree_interval.R
\name{tune.tree.interval}
\alias{tune.tree.interval}
\title{Wrapper function for tree-lasso objects.}
\usage{
tune.tree.interval(parms, x, y, lambda = NULL, nfolds = 5,
  foldid = NULL, num.nonpen = 0, seed = 12345, intercept = TRUE,
  standardize.response = FALSE, p = NULL, verbose = FALSE,
  parallel = FALSE, search.path = FALSE, threshold = threshold, ...)
}
\arguments{
\item{parms}{tuning parameter alpha for the tree-lasso object.}

\item{x, y}{\code{x} is a matrix where each row refers to a sample a each column refers to a gene; \code{y} is a factor which includes the class for each sample}

\item{lambda}{A user supplied lambda sequence.}

\item{nfolds}{number of cross-validation's folds, default 5.}

\item{foldid}{an optional vector of values between 1 and nfold identifying what fold each observation is in. If supplied, nfold can be missing.}

\item{num.nonpen}{number of predictors forced to be estimated (i.e., nonpenalization).}

\item{seed}{random seed}

\item{intercept}{should  intercept(s) be fitted (default=\code{TRUE}) or set to zero (\code{FALSE}).}

\item{standardize.response}{standardization for the response variables. Default: \code{TRUE}.}

\item{p}{the number of predictors from different data source.}

\item{verbose}{print the middle search information, default is \code{TRUE}.}

\item{parallel}{If \code{TRUE}, use parallel foreach to fit each lambda. If \code{c(TRUE,TRUE)}, use parallel foreach to fit each lambda and each fold.}

\item{search.path}{save the visited points, default is \code{FALSE}.}

\item{threshold}{threshold for estimated coefficients of the tree-lasso models.}
}
\description{
Wrapper function for tree-lasso objects used by epsgo function. This function is mainly used within the function \code{epsgo}.
}
\details{
IPFStructPenalty
}
