% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_inter_search.R
\name{summary.intsearch}
\alias{summary.intsearch}
\title{Summary results}
\usage{
\method{summary}{intsearch}(object, digits = max(3, getOption("digits") -
  3), verbose = TRUE, first.n = 5, ...)
}
\description{
Summarize the results from the oject of \code{epsgo}.
}
\details{
IPFStructPenalty
}
